benefits restructured as a function
